aruba_billing/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.4
django-allauth==0.55.2
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
gTTS==2.3.2
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
playsound==1.3.0
pycparser==2.21
pydub==0.25.1
PyJWT==2.8.0
python3-openid==3.2.0
pyttsx3==2.90
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
